+++ inbm/Changelog.md
@@ -3,13 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## X.X.X - YYYY-MM-DD
+ - (NEX-15262) Returns failure reason that matches the format required by MM
+
 ## 4.2.8 - 2024-11-29
 ### Changed
  - (NEXMANAGE-513) Record the granular log in download-only mode
  - (NEXMANAGE-1022) Adds Power Capabilities to Static Telemetry data.  Sends Static Telemetry data to INBS via the INBS Cloud Client in CloudAdapter.
    
 ### Added
-- (NEXMANAGE-826) Update granular log to support architecture independent package
+ - (NEXMANAGE-826) Update granular log to support architecture independent package
 
 ## 4.2.7 - 2024-11-8
 ### Changed